feat: phpunit test infrastructure for bundles api
