+++ api/test/.gitignore
@@ -1,0 +1,1 @@
+.phpunit.result.cache
