+++ api/test/CanaryTest.php
@@ -1,0 +1,16 @@
+<?php
+  use PHPUnit\Framework\TestCase;
+
+  class CanaryTest extends TestCase {
+    public function testCanary() {
+      $stack = [];
+      $this->assertSame(0, count($stack));
+
+      array_push($stack, 'foo');
+      $this->assertSame('foo', $stack[count($stack)-1]);
+      $this->assertSame(1, count($stack));
+
+      $this->assertSame('foo', array_pop($stack));
+      $this->assertSame(0, count($stack));
+    }
+  }
